Extract class from test runner projects
