+++ tests/TestClasses/ConvertersTests/ConvertersTestsClassWithConstructor.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace XamlParserTests
+{
+    public class ConvertersTestsClassWithConstructor
+    {
+        public int Int;
+
+        public TimeSpan Converted;
+
+        public ConvertersTestsClassWithConstructor(int i, TimeSpan converted)
+        {
+            Int = i;
+            Converted = converted;
+        }
+    }
+}
