+++ tests/TestClasses/ConvertersTests/ConvertersTestsClassWithoutConverter.cs
@@ -1,0 +1,9 @@
+﻿namespace XamlParserTests
+{
+    public class ConvertersTestsClassWithoutConverter
+    {
+        public string Value { get; set; }
+
+        public override string ToString() => Value;
+    }
+}
