+++ tests/TestClasses/ConvertersTests/ConvertersTestsEnum.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace XamlParserTests
+{
+    [Flags]
+    public enum ConvertersTestsEnum
+    {
+        First = 1,
+        Second = 2,
+        Third = 4
+    }
+}
