+++ tests/TestClasses/CustomConvertedType.cs
@@ -1,0 +1,7 @@
+﻿namespace XamlParserTests
+{
+    public class CustomConvertedType
+    {
+        public string Value { get; set; }
+    }
+}
