+++ tests/TestClasses/DeferredContentTests/ConstantRootObjectProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace XamlParserTests
+{
+    public class ConstantRootObjectProvider : ITestRootObjectProvider
+    {
+        public object RootObject { get; set; }
+    }
+}
