+++ tests/TestClasses/DeferredContentTests/DeferredContentTestsClass.cs
@@ -1,0 +1,10 @@
+﻿namespace XamlParserTests
+{
+    public class DeferredContentTestsClass
+    {
+        [Content, DeferredContent]
+        public object DeferredContent { get; set; }
+
+        public object ObjectProperty { get; set; }
+    }
+}
