+++ tests/TestClasses/EnumerableContentClass.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace XamlParserTests
+{
+    public class EnumerableContentClass
+    {
+        [Content]
+        public IEnumerable<SimpleSubClass> Children { get; set; } = new List<SimpleSubClass>();
+    }
+}
