+++ tests/TestClasses/ExtensionValueHolder.cs
@@ -1,0 +1,7 @@
+﻿namespace XamlParserTests
+{
+    public class ExtensionValueHolder
+    {
+        public object Value { get; set; }
+    }
+}
