+++ tests/TestClasses/Helpers.cs
@@ -7,8 +7,7 @@ namespace XamlParserTests
 {
     public static class Helpers
     {
-
-        public static void StructDiff(object parsed, object expected) 
+        public static void StructDiff(object parsed, object expected)
             => StructDiff(parsed, expected, "{root}", true);
 
         static void StructDiff(object parsed, object expected, string path, bool isRoot = false)
@@ -18,7 +17,7 @@ namespace XamlParserTests
             if ((parsed == null && expected != null) || (parsed != null && expected == null))
                 throw new Exception(
                     $"{path}: Null mismatch: {(parsed == null ? "null" : "not-null")}  {(expected == null ? "null" : "not-null")}");
-            
+
             if (parsed.GetType() != expected.GetType())
                 throw new Exception($"{path}: Type mismatch: {parsed.GetType()} {expected.GetType()}");
 
@@ -29,7 +28,7 @@ namespace XamlParserTests
             }
             else if (parsed is IDictionary dic)
             {
-                var dic2 = (IDictionary) expected;
+                var dic2 = (IDictionary)expected;
                 if (dic.Count != dic2.Count)
                     throw new Exception($"{path}: Dictionary count mismatch: {dic.Count} {dic2.Count}");
 
@@ -42,7 +41,7 @@ namespace XamlParserTests
             }
             else if (parsed is IList col)
             {
-                var col2 = (IList) expected;
+                var col2 = (IList)expected;
                 if (col.Count != col2.Count)
                     throw new Exception($"{path}: Collection count mismatch: {col.Count} {col2.Count}");
                 for (var c = 0; c < col.Count; c++)
@@ -56,7 +55,7 @@ namespace XamlParserTests
                     if (prop.DeclaringType == typeof(XamlAstNode)
                         && (prop.Name == "Line" || prop.Name == "Position"))
                     {
-                        if(!isRoot && (int)prop.GetValue(parsed) == 0)
+                        if (!isRoot && (int)prop.GetValue(parsed) == 0)
                             throw new Exception($"{path}.{prop.Name}: Missing line info (first)");
                         continue;
                     }
@@ -66,8 +65,8 @@ namespace XamlParserTests
             }
         }
 
-        public static T GetService<T>(this IServiceProvider prov) => (T) prov.GetService(typeof(T));
+        public static T GetService<T>(this IServiceProvider prov) => (T)prov.GetService(typeof(T));
 
-        
+
     }
 }
