+++ tests/TestClasses/InitializationTests/EventConstants.cs
@@ -1,0 +1,10 @@
+﻿namespace XamlParserTests
+{
+    public static class EventConstants
+    {
+        public static readonly string BeginInitEvent = "BeginInit";
+        public static readonly string EndInitEvent = "EndInit";
+        public static readonly string PropertySetEvent = "PropertySet";
+        public static readonly string ChildAddedEvent = "ChildAdded";
+    }
+}
