+++ tests/TestClasses/InitializationTests/InitializationTestsSupportInitializeClass.cs
@@ -1,0 +1,17 @@
+﻿using System.ComponentModel;
+
+namespace XamlParserTests
+{
+    public class InitializationTestsSupportInitializeClass : InitializationTestsClass, ISupportInitialize
+    {
+        public void BeginInit()
+        {
+            AddEvent(EventConstants.BeginInitEvent);
+        }
+
+        public void EndInit()
+        {
+            AddEvent(EventConstants.EndInitEvent);
+        }
+    }
+}
