+++ tests/TestClasses/InitializationTests/InitializationTestsTopDownClass.cs
@@ -1,0 +1,7 @@
+﻿namespace XamlParserTests
+{
+    [UsableDuringInitialization(true)]
+    public class InitializationTestsTopDownClass : InitializationTestsSupportInitializeClass
+    {
+    }
+}
