+++ tests/TestClasses/IntrinsicsTests/IntrinsicsTestsEnum.cs
@@ -1,0 +1,7 @@
+﻿namespace XamlParserTests
+{
+    public enum IntrinsicsTestsEnum : long
+    {
+        Foo = 100500
+    }
+}
