+++ tests/TestClasses/MarkupExtensionTests/MarkupExtensionContentDictionaryClass.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace XamlParserTests
+{
+    public class MarkupExtensionContentDictionaryClass
+    {
+        [Content]
+        public Dictionary<string, int> IntDic { get; } = new Dictionary<string, int>();
+    }
+}
