+++ tests/TestClasses/MarkupExtensionTests/ObjectTestExtension.cs
@@ -1,0 +1,12 @@
+﻿namespace XamlParserTests
+{
+    public class ObjectTestExtension
+    {
+        public object Returned { get; set; }
+
+        public object ProvideValue()
+        {
+            return Returned;
+        }
+    }
+}
