+++ tests/TestClasses/ServiceProviderTests/ServiceProviderTestsClass.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace XamlParserTests
+{
+    public class ServiceProviderTestsClass
+    {
+        public string Id { get; set; }
+        public object Property { get; set; }
+        public ServiceProviderTestsClass Child { get; set; }
+        [Content]
+        public List<ServiceProviderTestsClass> Children { get; } = new List<ServiceProviderTestsClass>();
+    }
+}
