+++ tests/TestClasses/ServiceProviderTests/ServiceProviderTestsClassForAttached.cs
@@ -1,0 +1,9 @@
+﻿namespace XamlParserTests
+{
+    public class ServiceProviderTestsClassForAttached
+    {
+        public static void SetAttachedProperty(ServiceProviderTestsClass target, string value)
+        {
+        }
+    }
+}
