+++ tests/TestClasses/ServiceProviderTests/ServiceProviderValueExtension.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace XamlParserTests
+{
+    public class ServiceProviderValueExtension
+    {
+        public object ProvideValue(IServiceProvider sp) =>
+            ((ExtensionValueHolder)sp.GetService(typeof(ExtensionValueHolder))).Value;
+    }
+}
