+++ tests/TestClasses/SimpleSubClass.cs
@@ -1,0 +1,7 @@
+﻿namespace XamlParserTests
+{
+    public class SimpleSubClass
+    {
+        public string Test { get; set; }
+    }
+}
