+++ tests/TestClasses/TestXamlLanguage.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using XamlParserTests;
 
 [assembly: XmlnsDefinition("test", "XamlParserTests")]
@@ -7,14 +6,12 @@ namespace XamlParserTests
 {
     public class ContentAttribute : Attribute
     {
-        
     }
 
     public class XmlnsDefinitionAttribute : Attribute
     {
         public XmlnsDefinitionAttribute(string xmlNamespace, string clrNamespace)
         {
-            
         }
     }
 
@@ -22,15 +19,13 @@ namespace XamlParserTests
     {
         public UsableDuringInitializationAttribute(bool usable)
         {
-            
         }
     }
 
     public class DeferredContentAttribute : Attribute
     {
-        
     }
-    
+
     public interface ITestRootObjectProvider
     {
         object RootObject { get; }
@@ -41,10 +36,9 @@ namespace XamlParserTests
         object TargetObject { get; }
         object TargetProperty { get; }
     }
-    
+
     public interface ITestUriContext
     {
         Uri BaseUri { get; set; }
     }
-    
 }
