+++ tests/XamlParserTests/ConvertersTests.cs
@@ -5,89 +5,6 @@ using Xunit;
 
 namespace XamlParserTests
 {
-    public class ConvertersTestClass
-    {
-        [Content]
-        public object ContentProperty { get; set; }
-        public long Int64Property { get; set; }
-        public bool BoolProperty { get; set; }
-        public double DoubleProperty { get; set; }
-        public float FloatProperty { get; set; }
-        public TimeSpan TimeSpanProperty { get; set; }
-        public Type TypeProperty { get; set; }
-        public UriKind UriKindProperty { get; set; }
-        public ConvertersTestValueType CustomProperty { get; set; }
-        public ConvertersTestsClassWithConverter TypeWithConverterProperty { get; set; }
-        [TypeConverter(typeof(PropertyTestConverter))]
-        public ConvertersTestsClassWithoutConverter PropertyWithConverter { get; set; }
-        public ConvertersTestsEnum EnumProperty { get; set; }
-    }
-
-    [Flags]
-    public enum ConvertersTestsEnum
-    {
-        First = 1,
-        Second = 2,
-        Third = 4
-    }
-    
-    public struct ConvertersTestValueType
-    {
-        public string Value { get; set; }
-        public override string ToString() => Value;
-
-        public static ConvertersTestValueType Parse(string s, IFormatProvider prov)
-        {
-            Assert.NotNull(prov);
-            return new ConvertersTestValueType() {Value = s};
-        }
-    }
-
-    [TypeConverter(typeof(TestConverter))]
-    public class ConvertersTestsClassWithConverter
-    {
-        public string Value { get; set; }
-        public override string ToString() => Value;
-    }
-    
-    public class ConvertersTestsClassWithoutConverter
-    {
-        public string Value { get; set; }
-        public override string ToString() => Value;
-    }
-
-    public class ConvertersTestsClassWithConstructor
-    {
-        public int Int;
-        public TimeSpan Converted;
-
-        public ConvertersTestsClassWithConstructor(int i, TimeSpan converted)
-        {
-            Int = i;
-            Converted = converted;
-        }
-    }
-
-    public class TestConverter : TypeConverter
-    {
-        public override object ConvertFrom(ITypeDescriptorContext context, CultureInfo culture, object value)
-        {
-            Assert.Equal(CultureInfo.InvariantCulture, culture);
-            Assert.NotNull(context.GetService<ITestRootObjectProvider>().RootObject);
-            return new ConvertersTestsClassWithConverter { Value = (string)value };
-        }
-    }
-
-    public class PropertyTestConverter : TypeConverter
-    {
-        public override object ConvertFrom(ITypeDescriptorContext context, CultureInfo culture, object value)
-        {
-            Assert.Equal(CultureInfo.InvariantCulture, culture);
-            Assert.NotNull(context.GetService<ITestRootObjectProvider>().RootObject);
-            return new ConvertersTestsClassWithoutConverter { Value = (string)value };
-        }
-    }
-
     public class ConvertersTests : CompilerTestBase
     {
