+++ tests/XamlParserTests/DictionaryTests.cs
@@ -1,18 +1,7 @@
-using System.Collections.Generic;
 using Xunit;
 
 namespace XamlParserTests
 {
-    public class SimpleClassWithDictionaryContent
-    {
-        public string Test { get; set; }
-        [Content]
-        public Dictionary<object, object> Children { get; set; } = new Dictionary<object, object>();
-        
-        public Dictionary<object, object> NonContentChildren { get; set; } = new Dictionary<object, object>();
-    }
-    
-    
     public class DictionaryTests : CompilerTestBase
     {
         [Fact]
